@@ -1,76 +1,82 @@
 {
-  "name": "gimler/symfony-rest-edition",
-  "license": "MIT",
-  "type": "project",
-  "description": "The \"Symfony REST Edition\" distribution",
-  "autoload": {
-    "psr-0": { "": "src/" }
-  },
-  "require-dev": {
-    "phpunit/phpunit": "4.4.*"
-  },
-  "require": {
-    "php": ">=5.3.3",
-    "symfony/symfony": "~2.3",
-    "symfony/icu": "1.0.*",
-    "doctrine/orm": "~2.2,>=2.2.3",
-    "doctrine/doctrine-bundle": "1.2.*",
-    "twig/extensions": "1.0.*",
-    "symfony/assetic-bundle": "2.3.*",
-    "symfony/swiftmailer-bundle": "2.3.*",
-    "symfony/monolog-bundle": "~2.3",
-    "sensio/distribution-bundle": "2.3.*",
-    "sensio/framework-extra-bundle": "2.3.*",
-    "sensio/generator-bundle": "2.3.*",
-    "jms/serializer-bundle": "0.13.*@dev",
-    "friendsofsymfony/rest-bundle": "1.4.*",
-    "friendsofsymfony/http-cache-bundle": "~1.0",
-    "willdurand/hateoas-bundle": "1.0.*@dev",
-    "hautelook/templated-uri-bundle": "~2.0",
-    "willdurand/rest-extra-bundle": "~1.0",
-    "incenteev/composer-parameter-handler": "~2.0",
-    "bomo/ical-bundle": "~1.0",
-    "friendsofsymfony/user-bundle": "~2.0@dev",
-    "snc/redis-bundle": "1.1.x-dev",
-    "predis/predis": "0.8.x-dev",
-    "jms/job-queue-bundle": "v1.2.1",
-    "pagerfanta/pagerfanta": "1.0.x"
-  },
-  "conflict": {
-    "symfony/routing": "2.4.1"
-  },
-  "scripts": {
-    "post-install-cmd": [
-      "Incenteev\\ParameterHandler\\ScriptHandler::buildParameters",
-      "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::clearCache",
-      "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::installAssets",
-      "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::installRequirementsFile"
-    ],
-    "post-update-cmd": [
-      "Incenteev\\ParameterHandler\\ScriptHandler::buildParameters",
-      "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::clearCache",
-      "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::installAssets",
-      "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::installRequirementsFile"
-    ]
-  },
-  "config": {
-    "bin-dir": "bin"
-  },
-  "minimum-stability": "stable",
-  "extra": {
-    "symfony-app-dir": "app",
-    "symfony-web-dir": "web",
-    "branch-alias": {
-      "dev-master": "2.3-dev"
+    "name": "troelsugiltjensen/koba",
+    "license": "proprietary",
+    "type": "project",
+    "autoload": {
+        "psr-4": {
+            "Itk\\ExchangeBundle\\": "src/Itk/ExchangeBundle",
+            "Koba\\AdminBundle\\": "src/Koba/AdminBundle",
+            "Koba\\ApiBundle\\": "src/Koba/ApiBundle",
+            "Koba\\MainBundle\\": "src/Koba/MainBundle"
+        },
+        "classmap": [
+            "app/AppKernel.php",
+            "app/AppCache.php"
+        ]
     },
-    "incenteev-parameters": {
-      "file": "app/config/parameters.yml"
+    "autoload-dev": {
+        "psr-4": {
+            "Tests\\": "tests/"
+        },
+        "files": [
+            "vendor/symfony/symfony/src/Symfony/Component/VarDumper/Resources/functions/dump.php"
+        ]
+    },
+    "require": {
+        "php": ">=5.5.9",
+        "bomo/ical-bundle": "^1.0",
+        "doctrine/doctrine-bundle": "^1.6",
+        "doctrine/doctrine-migrations-bundle": "^1.3",
+        "doctrine/orm": "^2.5",
+        "friendsofsymfony/rest-bundle": "^2.3",
+        "friendsofsymfony/user-bundle": "^2.1",
+        "incenteev/composer-parameter-handler": "^2.0",
+        "jms/job-queue-bundle": "^1.4",
+        "jms/serializer-bundle": "^2.4",
+        "predis/predis": "^1.1",
+        "sensio/distribution-bundle": "^5.0.19",
+        "sensio/framework-extra-bundle": "^5.0.0",
+        "snc/redis-bundle": "^2.1",
+        "symfony/monolog-bundle": "^3.1.0",
+        "symfony/polyfill-apcu": "^1.0",
+        "symfony/swiftmailer-bundle": "^2.6.4",
+        "symfony/symfony": "3.4.*",
+        "symfony/templating": "^3.4",
+        "twig/twig": "^1.0||^2.0"
+    },
+    "require-dev": {
+        "sensio/generator-bundle": "^3.0",
+        "symfony/phpunit-bridge": "^3.0"
+    },
+    "scripts": {
+        "symfony-scripts": [
+            "Incenteev\\ParameterHandler\\ScriptHandler::buildParameters",
+            "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::buildBootstrap",
+            "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::clearCache",
+            "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::installAssets",
+            "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::installRequirementsFile",
+            "Sensio\\Bundle\\DistributionBundle\\Composer\\ScriptHandler::prepareDeploymentTarget"
+        ],
+        "post-install-cmd": [
+            "@symfony-scripts"
+        ],
+        "post-update-cmd": [
+            "@symfony-scripts"
+        ]
+    },
+    "config": {
+        "sort-packages": true
+    },
+    "extra": {
+        "symfony-app-dir": "app",
+        "symfony-bin-dir": "bin",
+        "symfony-var-dir": "var",
+        "symfony-web-dir": "web",
+        "symfony-tests-dir": "tests",
+        "symfony-assets-install": "relative",
+        "incenteev-parameters": {
+            "file": "app/config/parameters.yml"
+        },
+        "branch-alias": null
     }
-  },
-  "repositories": [
-    {
-      "type": "vcs",
-      "url": "https://github.com/KOBADK/JMSJobQueueBundle"
-    }
-  ]
 }
